fix(scripts): drop the entrypoint guard from check-bundle
Review flagged the `import.meta.url === \`file://${process.argv[1]}\`` guard as
never matching. The stated reason was wrong — Node resolves argv[1] to an
absolute path, so it matched and the check did run. But the comparison is
fragile for other reasons: it breaks on percent-encoded paths (a space or any
non-ASCII character in the checkout path), on Windows separators, and through
symlinked checkouts. Each failure skips the check while still exiting 0 —
silently disabling the guard, which is the exact failure mode it exists to
prevent.

Rather than write a more careful comparison, remove the need for one. The pure
matchers move to bundle-collisions.mjs, which the tests import; check-bundle.mjs
is now CLI-only and always runs.

Re-verified: reintroducing the self-reference exits 1, the fixed tree exits 0.
